Improve download error handling

+++ ctapipe/utils/download.py
@@ -31,6 +31,7 @@ def download_file(url, path, auth=None, chunk_size=10240, progress=False):
     log.info(f"Downloading {url} to {path}")
     name = urlparse(url).path.split("/")[-1]
     path = Path(path)
+    part_file = None
 
     with requests.get(url, stream=True, auth=auth, timeout=5) as r:
         # make sure the request is successful
@@ -55,9 +56,9 @@ def download_file(url, path, auth=None, chunk_size=10240, progress=False):
                 for chunk in r.iter_content(chunk_size=chunk_size):
                     f.write(chunk)
                     pbar.update(len(chunk))
-        except:  # noqa we really want to catch everythin here
+        except BaseException:  # we really want to catch everything here
             # cleanup part file if something goes wrong
-            if part_file.is_file():
+            if part_file is not None and part_file.is_file():
                 part_file.unlink()
             raise
